// src/types/Producto.ts
export type Producto = {
  id: number;
  nombre: string;
  calificacion: number;
  marca: string;
  codigo_color: string;
  descripcion: string;
  precio: number;
  stock: number;
  fibra: string;
  grosor: string;
  peso: number;
  largo: number;
  calibre: number;
  agujas_sugeridas: number;
  ganchos_sugeridos: number;
  porcentaje_descuento: number;
  imagen_dir: string;
};